backend/fastapi-project/pyproject.toml
[tool.poetry]
name = "fastapi-project"
version = "0.1.0"
description = ""
authors = ["Admin <admin@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
python-multipart = "^0.0.6"
email-validator = "^2.1.0.post1"
passlib = {extras = ["bcrypt"], version = "^1.6.5"}
tenacity = "^8.2.3"
pydantic = "^2.5.3"
emails = "^0.6"
gunicorn = "^21.2.0"
jinja2 = "^3.1.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
httpx = "^0.25.1"
psycopg2 = "^2.9.9"
casbin = "^1.36.0"
casbin-sqlalchemy-adapter = "^0.5.2"
apscheduler = "^3.10.4"
itsdangerous = "^2.1.2"
fastapi-mail = "^1.4.1"
python-json-logger = "^2.0.7"
uvicorn = "^0.24.0.post1"

[tool.poetry.group.dev.dependencies]
mypy = "^1.7.0"
black = "^23.11.0"
isort = "^5.12.0"
autoflake = "^2.2.1"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
python-dotenv = "^1.0.1"


[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
